Created a components directory and extracted TopNavBar
