+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/vince/WebProjects/vrickey123.github.io/src/App.js":"1","/home/vince/WebProjects/vrickey123.github.io/src/index.js":"2"},{"size":680,"mtime":1606537919782,"results":"3","hashOfConfig":"4"},{"size":264,"mtime":1606536238271,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1qibd46",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/vince/WebProjects/vrickey123.github.io/src/App.js",[],"/home/vince/WebProjects/vrickey123.github.io/src/index.js",[]]
+[{"/home/vince/WebProjects/vrickey123.github.io/src/App.js":"1","/home/vince/WebProjects/vrickey123.github.io/src/index.js":"2"},{"size":295,"mtime":1606538576886,"results":"3","hashOfConfig":"4"},{"size":264,"mtime":1606536238271,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1qibd46",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/vince/WebProjects/vrickey123.github.io/src/App.js",[],"/home/vince/WebProjects/vrickey123.github.io/src/index.js",[]]
